Local_backend/backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
sqlalchemy
python-dotenv
passlib[bcrypt]
pyJWT
pydantic
pydantic[email]
email-validator